+++ src/components/Email/Email.tsx
@@ -1,9 +1,26 @@
 import { ReactNode } from 'react';
+import { BaseStyleProp } from '../types';
+import { makeStyles } from '../../utils/makeStyles';
 
-export interface EmailProps {
+type EmailStyles = 'root';
+
+export interface EmailProps extends BaseStyleProp<EmailStyles> {
   children?: ReactNode;
 }
 
-export const Email = ({ children }: EmailProps): JSX.Element => {
-  return <div>{children}</div>;
+const useStyles = makeStyles({
+  root: {
+    margin: '0px auto',
+    maxWidth: '600px',
+  },
+});
+
+export const Email = ({ children, className, classes }: EmailProps): JSX.Element => {
+  const styles = useStyles({ classes });
+
+  return (
+    <div style={styles.root} className={className}>
+      {children}
+    </div>
+  );
 };