+++ src/components/types/index.ts
@@ -1,0 +1,6 @@
+import { CSSProperties } from 'react';
+
+export interface BaseStyleProp<T extends string> {
+  classes?: Partial<Record<T, CSSProperties>>;
+  className?: string;
+}